Modified to run agents in own environment, indep
Agents running with the same environment may have an effect on each others performance. As this is intended as a demonstration of Thompson sampling vs Epsilon sampling, this is not beneficial as it may provide bias during initial random stages.
The convnetjs code is now more encapsulated; however, certain elements are still hard-coded. A better mechanism for scaling the canvases needs to be implemented which is dynamic, likewise hard-coding Id's is not a scalable solution. It would be nice to run multiple selection strategies with a direct compare mechanism.